feat: align v2 data imports and shop/skin decoding

+++ internal/answer/shopping_command_answer.go
@@ -41,7 +41,7 @@ func ShoppingCommandAnswer(buffer *[]byte, client *connection.Client) (int, int,
 		Result: proto.Uint32(0),
 	}
 
-	if !client.Commander.HasEnoughResource(shopOffer.ResourceID, shopOffer.ResourceNumber) {
+	if !client.Commander.HasEnoughResource(shopOffer.ResourceID, uint32(shopOffer.ResourceNumber)) {
 		logger.LogEvent("Shop", "Purchase", fmt.Sprintf("uid=%d does not have enough resources", client.Commander.CommanderID), logger.LOG_LEVEL_INFO)
 		response.Result = proto.Uint32(1)
 		return 0, 16002, nil
@@ -52,20 +52,20 @@ func ShoppingCommandAnswer(buffer *[]byte, client *connection.Client) (int, int,
 	switch shopOffer.Type {
 	case 1: // bought resources
 		for i, resourceId := range shopOffer.Effects {
-			client.Commander.AddResource(uint32(resourceId), shopOffer.Number)
+			client.Commander.AddResource(uint32(resourceId), uint32(shopOffer.Number))
 			response.DropList[i] = &protobuf.DROPINFO{
 				Type:   proto.Uint32(shopOffer.Type), // ressource
 				Id:     proto.Uint32(uint32(resourceId)),
-				Number: proto.Uint32(shopOffer.Number),
+				Number: proto.Uint32(uint32(shopOffer.Number)),
 			}
 		}
 	case 2: // packs
 		for i, packId := range shopOffer.Effects {
-			client.Commander.AddItem(uint32(packId), shopOffer.Number)
+			client.Commander.AddItem(uint32(packId), uint32(shopOffer.Number))
 			response.DropList[i] = &protobuf.DROPINFO{
 				Type:   proto.Uint32(shopOffer.Type), // item
 				Id:     proto.Uint32(uint32(packId)),
-				Number: proto.Uint32(shopOffer.Number),
+				Number: proto.Uint32(uint32(shopOffer.Number)),
 			}
 		}
 	case 4: // merit shop, to implement
@@ -76,7 +76,7 @@ func ShoppingCommandAnswer(buffer *[]byte, client *connection.Client) (int, int,
 			response.DropList[i] = &protobuf.DROPINFO{
 				Type:   proto.Uint32(shopOffer.Type), // skin
 				Id:     proto.Uint32(uint32(skinId)),
-				Number: proto.Uint32(shopOffer.Number),
+				Number: proto.Uint32(uint32(shopOffer.Number)),
 			}
 		}
 	case 12: // operation siren, to implement
@@ -90,6 +90,6 @@ func ShoppingCommandAnswer(buffer *[]byte, client *connection.Client) (int, int,
 		logger.LogEvent("Shop", "Purchase", fmt.Sprintf("uid=%d bought #%d successfully!", client.Commander.CommanderID, shopOffer.ID), logger.LOG_LEVEL_INFO)
 	}
 
-	client.Commander.ConsumeResource(shopOffer.ResourceID, shopOffer.ResourceNumber)
+	client.Commander.ConsumeResource(shopOffer.ResourceID, uint32(shopOffer.ResourceNumber))
 	return client.SendMessage(16002, &response)
 }
